Rename DNO.
