+++ src/continuity/operators/__init__.py
@@ -15,7 +15,7 @@ from .operator import Operator
 from .neuraloperator import NeuralOperator
 from .deeponet import DeepONet
 from .belnet import BelNet
-from .deep_neural_operator import DeepNeuralOperator
+from .dno import DeepNeuralOperator
 from .shape import OperatorShapes
 
 __all__ = [
