+++ tests/operators/test_dno.py
